Improve usability by removing autovcpkg-build crate. Add vgtk example.
The autovcpkg can be both a dependency or/and a build-dependency.
When used as a dependency, it will act as a -sys crate compiling the
native libraries.
When used as a build-dependency, it can then be used in a build.rs to
instruct cargo on how to find and link to native libraries.